Separate server to allow for simulator
